% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallSciensanoShinyTemplate.R
\name{CopyShinyDashboard}
\alias{CopyShinyDashboard}
\title{Create Shinydashboard template directory}
\usage{
CopyShinyDashboard(destination = ".")
}
\arguments{
\item{destination}{A character string specifying the directory where the template
should be copied.}
}
\description{
Copy Shinydashboard Template
}
\details{
This function copies the Shinydashboard template from the package to a specified directory
for further modification.
}
