% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallSciensanoShinyTemplate.R
\name{CopyShinyTemplate}
\alias{CopyShinyTemplate}
\title{Create shiny app template directory}
\usage{
CopyShinyTemplate(destination = ".")
}
\arguments{
\item{destination}{A character string specifying the directory where the template
should be copied.}
}
\description{
Copy Shiny App Template
}
\details{
This function copies the Shiny app template from the package to a specified directory
for further modification.
}
