% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallSciensanoShinyTemplate.R
\name{runTemplateDashboard}
\alias{runTemplateDashboard}
\title{Run the Custom Shinydashboard Template}
\usage{
runTemplateDashboard()
}
\description{
This function launches the custom Shinydashboard template included in the package.
}
