% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallSciensanoShinyTemplate.R
\name{runTemplateShiny}
\alias{runTemplateShiny}
\title{Run the Custom Shiny App Template}
\usage{
runTemplateShiny()
}
\description{
This function launches the custom Shiny app template included in the package.
}
