% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scienplt.R
\name{sciensano_plot}
\alias{sciensano_plot}
\title{This function helps you create a Sciensano styled plot easily}
\usage{
sciensano_plot(
  x = "data",
  ColourPal = c("#006633", "#3aaa35", "#bccf00", "#fad500", "#f49e24", "#c85927",
    "#e190bd", "#b33992", "#563b8c", "#769fd3", "#57bfb7", "#298691"),
  Inverse = F
)
}
\arguments{
\item{x}{a ggplot object}

\item{ColourPal}{set1, set2 or set3}

\item{Inverse}{Inverted colors (T or F)}
}
\value{
a Sciensano styled ggplot
}
\description{
This function helps you create a Sciensano styled plot easily
}
\examples{
library(ggplot2)

ggplot_object <- ggplot(mtcars) +
  aes(disp, mpg, color = cyl) +
  geom_point()

sciensano_plot(ggplot_object)

}
