% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scientbl.R
\name{sciensano_table}
\alias{sciensano_table}
\title{sciensano_table() is using the flextable package to style your tables.
It can be used within any type of Rmarkdown.}
\usage{
sciensano_table(x, caption = NULL, autofit = F, scroll = T)
}
\arguments{
\item{x}{a table}

\item{caption}{a caption}

\item{autofit}{default to FALSE, set to TRUE for page fitting tables in Word Rmd.}

\item{scroll}{default to TRUE, adds a scroll for tables bigger than 20 rows (only in HTLM format).}
}
\value{
a Sciensano styled table
}
\description{
sciensano_table() is using the flextable package to style your tables.
It can be used within any type of Rmarkdown.
}
\examples{
sciensano_table(head(mtcars))

}
